Training material
Training material for testing the database with QGIS. This material was a base for testing if the database model for general land use plan created in Tampere had all the necessary information and also if the database is maintained for QGIS what additional tools it needs. The database may change in the future so this material might not be valid after 10.9.2018.